Added the MK45 Hammer
The Mk45 Hammer serves as an immobile, mid-tier option of NATO artillery. While powerful with a slew of ammunition types, the Hammer's immobility and range akin to the M4 Scorcher make it a fair bit vulnerable. However it makes up for it with the ability to be controlled by a UAV Operator from afar, and its cheap cost compared to mobile artillery pieces.