Added Triton PMC Preset
'Triton' is a Eastern-backed PMC notorious for its tendency to work for the highest bidder. While primarily sporting Russian gear, hundreds of elite operatives apart of Triton, known as 'Ravens', are given the best gear, and tasked with waging proxy wars on behalf of their benefactors.

Although the PMC holds a pool of armored vehicles and armed attack helicopters, they are incredibly expensive to requisition, and as such the PMC often finds itself using 'hit-and-run' strike tactics, with cheap LSV's and helicopters at their disposal.

In the ongoing international conflict, there is little room for hesitation, and a constantly growing demand for contractors who find the hazard pay more than worthwhile.